requirements.txt
asgiref==3.9.1
Django==5.2.5
django-environ==0.12.0
django-simple-history==3.10.1
djangorestframework==3.16.1
mysqlclient==2.2.7
sqlparse==0.5.3
typing_extensions==4.14.1
